LayoutTests/storage/indexeddb/database-odd-names-expected.txt

Test IndexedDB odd value datatypes

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;
PASS Boolean(indexedDB && IDBCursor && IDBDatabase && IDBDatabaseError && IDBDatabaseException && IDBFactory && IDBIndex && IDBKeyRange && IDBObjectStore && IDBRequest && IDBTransaction) is true

opening a database named null
indexedDB.open(testData[nextToOpen].name, testData[nextToOpen].description)
opening a database named faihu
indexedDB.open(testData[nextToOpen].name, testData[nextToOpen].description)
opening a database named unpaired surrogate
indexedDB.open(testData[nextToOpen].name, testData[nextToOpen].description)
opening a database named fffe
indexedDB.open(testData[nextToOpen].name, testData[nextToOpen].description)
opening a database named ffff
indexedDB.open(testData[nextToOpen].name, testData[nextToOpen].description)
opening a database named line separator
indexedDB.open(testData[nextToOpen].name, testData[nextToOpen].description)
PASS successfullyParsed is true

TEST COMPLETE
